Add files via upload
This is a skeleton version of an automated VBA checklist I created for my prior workplace to help our internal auditor with internal control requirements. This checklist was manual prior to the creation of this automated checklist and it helped save time, save paper, was user-friendly, and added additional features which enforced segregation of duties. The checklist functioned by the preparer first going into the file and filling out their checkmarks/NA for each fund preparation requirement and then simply pressing a button to lock their column of the spreadsheet and have their initials and timestamp recorded and locked. This process was repeated by the reviewer and final manager. This skeletal version is trimmed down from the original document and the code has been locked for proprietary reasons.